Add files via upload
Puzzle input is the code for putting in the sudoku puzzle and it spits it out in text form. The sudoku solver file is the code for solving an inputted sudoku puzzle. Ultimate Sudoku solver file is a combination of both so you can input the sudoku puzzle with a GUI and it will convert it to text for the sudoku solver to use